Ficha completa e tela

+++ app/controllers/encantamentos_controller.rb
@@ -1,6 +1,7 @@
 class EncantamentosController < ApplicationController
   before_action :set_personagem
   before_action :carregar_opcoes
+  before_action :set_encantamento, only: [:edit, :update]
 
   def new
     @encantamento = novo_encantamento
@@ -10,22 +11,22 @@ class EncantamentosController < ApplicationController
 
   def create
     @encantamento = novo_encantamento(encantamento_params)
+    salvar_encantamento(notice: "Encantamento salvo com sucesso!", template: :new)
+  end
 
-    if @encantamento.save
-      modificador_ids_atuais.each_with_index do |mod_id, index|
-        @encantamento.encantamento_modificadors.create!(modificador_id: mod_id, ordem: index)
-      end
-      @encantamento.recalcular!
-      redirect_to destino_apos_salvar, notice: "Encantamento salvo com sucesso!"
-    else
-      @modificador_ids_selecionados = modificador_ids_atuais
-      @resultado = calcular_resultado
-      render :new, status: :unprocessable_entity
-    end
+  def edit
+    @modificador_ids_selecionados = @encantamento.modificadors.pluck(:id).map(&:to_s)
+    @resultado = resultado_para_encantamento(@encantamento)
+  end
+
+  def update
+    @encantamento.assign_attributes(encantamento_params)
+    salvar_encantamento(notice: "Encantamento atualizado com sucesso!", template: :edit)
   end
 
   def preview
-    @encantamento = novo_encantamento(encantamento_params)
+    @encantamento = params[:id] ? Encantamento.find(params[:id]) : novo_encantamento
+    @encantamento.assign_attributes(encantamento_params)
     @modificador_ids_selecionados = modificador_ids_atuais
     @resultado = calcular_resultado
     respond_to { |format| format.turbo_stream }
@@ -37,8 +38,27 @@ class EncantamentosController < ApplicationController
     @personagem = Personagem.find(params[:personagem_id]) if params[:personagem_id].present?
   end
 
+  def set_encantamento
+    @encantamento = Encantamento.find(params[:id])
+  end
+
   def novo_encantamento(attrs = {})
     @personagem ? @personagem.encantamentos.new(attrs) : Encantamento.new(attrs)
+  end
+
+  def salvar_encantamento(notice:, template:)
+    if @encantamento.save
+      @encantamento.encantamento_modificadors.destroy_all
+      modificador_ids_atuais.each_with_index do |mod_id, index|
+        @encantamento.encantamento_modificadors.create!(modificador_id: mod_id, ordem: index)
+      end
+      @encantamento.recalcular!
+      redirect_to destino_apos_salvar, notice: notice
+    else
+      @modificador_ids_selecionados = modificador_ids_atuais
+      @resultado = calcular_resultado
+      render template, status: :unprocessable_entity
+    end
   end
 
   def destino_apos_salvar
@@ -46,7 +66,11 @@ class EncantamentosController < ApplicationController
   end
 
   def preview_path_atual
-    @personagem ? preview_personagem_encantamentos_path(@personagem) : preview_encantamentos_path
+    if @encantamento&.persisted?
+      @personagem ? preview_personagem_encantamento_path(@personagem, @encantamento) : preview_encantamento_path(@encantamento)
+    else
+      @personagem ? preview_personagem_encantamentos_path(@personagem) : preview_encantamentos_path
+    end
   end
   helper_method :preview_path_atual
 
@@ -63,17 +87,13 @@ class EncantamentosController < ApplicationController
   end
 
   def encantamento_params
-    params.require(:encantamento).permit(:nome, :forma_id, :transmutacao_id)
+    params.require(:encantamento).permit(:nome, :forma_id, :transmutacao_id, :observacao)
   end
 
   def modificador_ids_atuais
     existentes = Array(params[:modificador_ids]).reject(&:blank?)
     existentes += [params[:novo_modificador_id]] if params[:novo_modificador_id].present?
-
-    if params[:remover_indice].present?
-      existentes.delete_at(params[:remover_indice].to_i)
-    end
-
+    existentes.delete_at(params[:remover_indice].to_i) if params[:remover_indice].present?
     existentes
   end
 
@@ -87,5 +107,12 @@ class EncantamentosController < ApplicationController
     modificadores = ids.map { |id| por_id[id] }.compact
 
     CalculoEncantamento.new(forma: forma, transmutacao: transmutacao, modificadores: modificadores).calcular
+  end
+
+  def resultado_para_encantamento(encantamento)
+    CalculoEncantamento.new(
+      forma: encantamento.forma, transmutacao: encantamento.transmutacao,
+      modificadores: encantamento.modificadors
+    ).calcular
   end
 end
